*** empty log message *** Original: 7eca9b0a25cb3d1a87eeea5f4488f8af7e67f0a0

+++ lalapps/src/pulsar/FDS_isolated/combiner.py
@@ -14,21 +14,25 @@ from optparse import OptionParser
 ## command line options
 parser = OptionParser()
 parser.add_option("-w", "--workdir", dest="workdir",
-                  help="Directory where the result file will be.",
+                  help="Directory where the result directory will be.",
 		  default="./")
 parser.add_option("-t", "--targetdir", dest="targetdir",
                   help="Directory under which the data files will be.",
-		  default="./targetdir/")
+                  default="./separate_targetdir/")
+parser.add_option("-i", "--tmpdir", dest="tmpdir",
+                  help="Directory where the result files will be "
+		  "under workdir.",
+		  default="separate_tmpdir/")
 parser.add_option("-r", "--resultfile", dest="resultfile",
                   help="Name of the result file.",
 		  default="Combined")
 (options, args) = parser.parse_args()
 
 
-
-## set-up directories and files
+## set-up and normalize directory names.
 workdir=os.path.abspath(options.workdir)+"/"
 targetdir=os.path.abspath(options.targetdir)+"/"
+tmpdir=os.path.abspath(workdir+options.tmpdir)+"/"
 resultfile=os.path.abspath(workdir+options.resultfile)
 ## check if there are such directories.
 if os.path.isdir(workdir) is False:
@@ -42,55 +46,72 @@ print "result file name: "+resultfile
 
 
 
+try: os.mkdir(tmpdir)
+except OSError, err:
+    import errno
+    print "Warning:", err
+    uinput=raw_input("Are you sue to delete "+tmpdir+".: [y/n]")
+    if uinput is "y":
+       os.system("rm -rf "+tmpdir)
+       os.mkdir(tmpdir)
+    else:
+       sys.exit()		   
+
+
 
 ##---------------------------------------------------------------
 ## Functions
 ##---------------------------------------------------------------
 
-def addfilename(targetfile="",niter=2):
-    """
-      Replace the line starting as %n (n=1,2,...,niter) with %filename_n
-    """
-    __tmpfilename=targetfile+"_tmp_addfilename"
-    for i in range(niter):
-        shutil.move(targetfile,__tmpfilename)
-        label=str(i+1)
-	targetbase=os.path.basename(targetfile)
-        __command=["sed -e s/%",label,"/%",
-		   targetbase, "_",label,"/ < ", __tmpfilename,
-                   " > ", targetfile]
-        __comexe=string.join(__command,"")
-        os.system(__comexe)
-	os.remove(__tmpfilename)
-
-
-def textcutter(targetfile="",textfrom="",textto=""):
+def addheaderinfo(targetfile="",info=""):
    """
-      From the targetfile, take contents between textfrom and textto.
-   """
-   __errmsg="Error"
-   if (targetfile is "" or textfrom is "" or textto is ""):
+	add info to the targetfile header. 
+   """	
+   __funcname="addheaderinfo"  
+   __errmsg="Error"+" in "+__funcname
+   if (targetfile is ""):
       print __errmsg, "Enough numbers of arguments are not given." 
       sys.exit()    
    if os.path.isfile(targetfile) is False:
-      print __errmsg, targetfile, "does not exit!!" 
+      print __errmsg, targetfile, "does not exist!!" 
       sys.exit() 
-   __tmpfilename=targetfile+"_tmp_textcutter"
+##      
+   __tmpfilename=targetfile+"_tmp_addfilename"
    shutil.move(targetfile,__tmpfilename)
-   __command=["sed -e /",
-              textfrom,
-	      "/,/",
-              textto,
-	      "/p -e d ", 
-              __tmpfilename,
-              " | grep -v ", 
-              textto, 
-              " > ",
-              targetfile]
+   __command=["sed -e '1i",
+              info, "' < ", __tmpfilename,
+              " > ", targetfile]
    __comexe=string.join(__command,"")
    os.system(__comexe)
    os.remove(__tmpfilename)
- 
+
+
+
+def addheaderinfo2(targetfile="",info="",headerlabel="%"):
+   """
+	add info to the targetfile header. The file header should be 
+	indicated by headerlabel. 
+   """	
+   __funcname="addheaderinfo"  
+   __errmsg="Error"+" in "+__funcname
+   if (targetfile is ""):
+      print __errmsg, "Enough numbers of arguments are not given." 
+      sys.exit()    
+   if os.path.isfile(targetfile) is False:
+      print __errmsg, targetfile, "does not exist!!" 
+      sys.exit() 
+##      
+   __tmpfilename=targetfile+"_tmp_addfilename"
+   shutil.move(targetfile,__tmpfilename)
+   __command=["sed -e s/",headerlabel,"/",headerlabel,
+              info, "___/ < ", __tmpfilename,
+              " > ", targetfile]
+   __comexe=string.join(__command,"")
+   os.system(__comexe)
+   os.remove(__tmpfilename)
+
+
+
 
 def addfooterinfo(targetfile="",info=""):
    """
@@ -115,17 +136,28 @@ def addfooterinfo(targetfile="",info=""):
    os.remove(__tmpfilename)
 
 
-
-def appendfile(fromfile="",tofile=sys.stdout):
+def addinfoToEachRaw(targetfile="",info=""):
     """
-       Append fromfile to tofile.
+       add info to each raw. 
     """
-    if fromfile is "":
-       print "Error"
-       sys.exit()	
-    os.system("cat "+fromfile+" >> "+tofile)
-
-
+    __funcname="addinfoToEachRaw"  
+    __errmsg="Error"+" in "+__funcname
+    if (targetfile is ""):
+        print __errmsg, "Enough numbers of arguments are not given." 
+        sys.exit()    
+    if os.path.isfile(targetfile) is False:
+        print __errmsg, targetfile, "does not exit!!" 
+        sys.exit() 
+##      
+    __tmpfilename=targetfile+"_tmp_addfilename"
+    shutil.move(targetfile,__tmpfilename)
+    __command=["sed -e 's/^/",
+               info, "    /' < ", __tmpfilename,
+               " > ", targetfile]
+    __comexe=string.join(__command,"")
+    os.system(__comexe)
+    os.remove(__tmpfilename)
+    
 
 def deleteinfo(targetfile="",label=""):
     """
@@ -150,47 +182,80 @@ def deleteinfo(targetfile="",label=""):
     os.remove(__tmpfilename)
 
 
-
-
-def addinfoToEachRaw(targetfile="",info=""):
-    """
-       add info to each raw. 
-    """
-    __funcname="addinfoToEachRaw"  
-    __errmsg="Error"+" in "+__funcname
-    if (targetfile is ""):
-        print __errmsg, "Enough numbers of arguments are not given." 
-        sys.exit()    
-    if os.path.isfile(targetfile) is False:
-        print __errmsg, targetfile, "does not exit!!" 
-        sys.exit() 
-##      
-    __tmpfilename=targetfile+"_tmp_addfilename"
-    shutil.move(targetfile,__tmpfilename)
-##    __command=["sed -e 's/$/   ",
-    __command=["sed -e 's/^/   ",
-               info, "/' < ", __tmpfilename,
-               " > ", targetfile]
-    __comexe=string.join(__command,"")
-    os.system(__comexe)
-    os.remove(__tmpfilename)
-
     
+
+
+
+def fileseparater(targetfile="",textto="",tmpdir="",niter=2,sectionlabel="%"):
+   """
+      split the sections in the targetfile into files and put them on 
+      tmpdir. The sections are defined by sectionlabel.
+   """
+   __funcname="addheaderinfo"  
+   __errmsg="Error"+" in "+__funcname
+   if (targetfile is "" or textto is "" or tmpdir is ""):
+      print __errmsg, "Enough numbers of arguments are not given." 
+      sys.exit()    
+   if os.path.isfile(targetfile) is False:
+      print __errmsg, targetfile, "does not exist!!" 
+      sys.exit() 
+##
+   __tmpfilename=targetfile+"_tmp_fileseparater"
+   shutil.move(targetfile,__tmpfilename)
+   for i in range(niter-1):
+       __textfrom=str(i+1)
+       __textto="%"+str(i+2)
+       __resultfile=targetfile+"___"+__textfrom
+       __command=["sed -e /",
+		  sectionlabel,     
+		  __textfrom,
+		  "/,/",
+		  __textto,
+		  "/p -e d ", 
+		  __tmpfilename,
+		  " | grep -v ", 
+		  __textto, 
+		  " > ",
+		  __resultfile]    
+       __comexe=string.join(__command,"")
+       os.system(__comexe)
+       shutil.move(__resultfile,tmpdir)
+   __textfrom=str(niter)
+   __resultfile=targetfile+"___"+__textfrom
+   __command=["sed -e /",
+              sectionlabel,
+              __textfrom,
+	      "/,/",
+              textto,
+	      "/p -e d ", 
+              __tmpfilename,
+              " | grep -v ", 
+              textto, 
+              " > ",
+              __resultfile]    
+   __comexe=string.join(__command,"")
+   os.system(__comexe)
+   shutil.move(__resultfile,tmpdir)
+   os.remove(__tmpfilename)
+ 
+
+
+def appendfile(fromfile="",tofile=sys.stdout):
+    """
+       Append fromfile to tofile.
+    """
+    if fromfile is "":
+       print "Error"
+       sys.exit()	
+    os.system("cat "+fromfile+" >> "+tofile)
+    
+
 ##---------------------------------------------------------------
 ## Ask intention
-if os.path.isfile(resultfile):		
+if os.path.isdir(tmpdir):		
    print "Current size of the result file."
    os.system("du -h "+resultfile)
-   uinput=raw_input("Do you want to delete [d], or append [a]?:[d/a]")
-   if uinput is "d":
-       os.system("rm -f "+resultfile)
-   elif uinput is "a":
-       pass
-   else:
-       sys.exit()
-       
-
-print "size of the target file before unzip"
+print "size of the target directory before unzip"
 os.system("du -h "+targetdir+" | tail -n 1")
 uinput=raw_input("Are you sure to proceed?:[y/n]")
 if uinput is "y":
@@ -201,7 +266,8 @@ else:
 
 ##---------------------------------------------------------------
 ## Main 
-fileid=1
+
+## This loop separates files.
 for parentdir, childdirs, files in os.walk(targetdir):
     for file in files:              # For each file under the targetdir,
 	if file is not []:          # If there is a file, 
@@ -211,19 +277,22 @@ for parentdir, childdirs, files in os.walk(targetdir):
 	   zipfilename=copiedfile+"_tmp.zip"   
 	   shutil.move(copiedfile,zipfilename)
 	   os.system("unzip -qa "+zipfilename)   # Unzip the copied file
-	   os.remove(zipfilename)
-           ## take sections from textfrom to textto.
-	   textcutter(targetfile=copiedfile,textfrom=textfrom,textto=textto)
-           ## delete line whose first line has label.
-           deleteinfo(targetfile=copiedfile,label="%")
-           addinfoToEachRaw(targetfile=copiedfile,info=str(fileid)+"   ")
-           fileid+=1
-
-           ## write the file name at each section header.
-##	   addfilename(targetfile=copiedfile)
-           ## append the file into one result file.
-	   appendfile(fromfile=copiedfile,tofile=resultfile)
-	   os.remove(copiedfile)
+	   os.remove(zipfilename)	
+	   fileseparater(targetfile=copiedfile,textto=textto,tmpdir=tmpdir)
+## This loop add various information to files.
+fileid=1
+for parentdir, childdirs, files in os.walk(tmpdir):
+    for file in files:              # For each file under the targetdir,
+	if file is not []:          # If there is a file, 
+	   targetfile=os.path.join(parentdir,file)
+	   deleteinfo(targetfile=targetfile,label="%")
+           addinfoToEachRaw(targetfile=targetfile,info=str(fileid))
+           fileid+=1           
+#	   addheaderinfo(targetfile=targetfile,info="%"+file)
+#	   addfooterinfo(targetfile=targetfile,info="%DONE")
+	   appendfile(fromfile=targetfile,tofile=resultfile)
 
 
 addfooterinfo(targetfile=resultfile,info="%DONE")
+
+
